+++ tchannel/testing/vcr/config.py
@@ -39,16 +39,21 @@ VCRProxyClient = client_for('vcr', VCRProxy)
 class _CassetteContext(object):
     """Lets use_cassette be used as a context manager and a decorator."""
 
-    def __init__(self, path, record_mode, inject):
+    def __init__(self, path, record_mode, inject, matchers):
         self.path = path
         self.record_mode = record_mode
         self.inject = inject
+        self.matchers = matchers
 
         self._exit_stack = contextlib2.ExitStack()
 
     def __enter__(self):
         cassette = self._exit_stack.enter_context(
-            Cassette(path=self.path, record_mode=self.record_mode)
+            Cassette(
+                path=self.path,
+                record_mode=self.record_mode,
+                matchers=self.matchers,
+            )
         )
 
         server = self._exit_stack.enter_context(
@@ -99,7 +104,7 @@ class _CassetteContext(object):
                 return function(*args, **kwargs)
 
 
-def use_cassette(path, record_mode=None, inject=False):
+def use_cassette(path, record_mode=None, inject=False, matchers=None):
     """Use or create a cassette to record/replay TChannel requests.
 
     This may be used as a context manager or a decorator.
@@ -141,8 +146,42 @@ def use_cassette(path, record_mode=None, inject=False):
         If True, when ``use_cassette`` is used as a decorator, the cassette
         object will be injected into the function call as the first argument.
         Defaults to False.
+    :param matchers:
+        Used to configure the request attributes which VCR matches on. This is
+        a list of request attributes or a function that accepts a list of
+        request attributes and returns a new list of attributes (this may be
+        used to transform the default list of matchers used by VCR).  A
+        recorded response is replayed if all specified attributes of the
+        corresponding request match the request that is being made. Valid
+        attirbutes are: ``serviceName``, ``hostPort``, ``endpoint``,
+        ``headers``, ``body``, ``argScheme``, and ``transportHeaders``.
+
+        For example,
+
+        .. code-block:: python
+
+            @vcr.use_cassette('tests/data/foo.yaml', matchers=['body']):
+            def test_foo():
+                # ...
+
+            def no_headers(matchers):
+                matchers.remove('headers')
+                return matchers
+
+            @vcr.use_cassette('tests/data/bar.yaml', matchers=no_headers):
+            def test_bar():
+                # ...
+
+        By default, VCR matches on the attributes: ``serviceName``,
+        ``endpoint``, ``headers``, ``body``, and ``argScheme``.
     """
-    return _CassetteContext(path=path, record_mode=record_mode, inject=inject)
+
+    return _CassetteContext(
+        path=path,
+        record_mode=record_mode,
+        inject=inject,
+        matchers=matchers,
+    )
 
     # TODO create some sort of configurable VCR object which implements
     # use_cassette. Top-level use_cassette can just use a default instance.